+++ src/Loss.cpp
@@ -8,7 +8,7 @@ Tensor MSELoss::operator()(const Tensor& pred, const Tensor& target) {
     sqError = error.pow(2.0);
     sumError = sqError.sum();
     
-    double N = (double)pred.size; 
+    float N = (float)pred.size; 
     
     Tensor outLoss({1});
     outLoss.data[0] = sumError.data[0] / N;
@@ -17,7 +17,7 @@ Tensor MSELoss::operator()(const Tensor& pred, const Tensor& target) {
     outLoss.prev.push_back(&sumError);
     
     Tensor* sumPtr = &sumError;
-    outLoss._backward = [N, sumPtr](const double* outGrad) {
+    outLoss._backward = [N, sumPtr](const float* outGrad) {
         sumPtr->grad[0] += (1.0 / N) * outGrad[0];
     };
